refactor(hosting): move listing page as first level of architecture
