+++ packages/manager/apps/web/client/app/hosting/dashboard/api-translator.service.js
