+++ packages/manager/apps/web/client/app/hosting/dashboard/hosting.constant.js
