+++ packages/manager/apps/web/client/app/hosting/dashboard/hosting.controller.js
