+++ packages/manager/apps/web/client/app/hosting/dashboard/hosting.routes.js
@@ -1,7 +1,7 @@
 import controller from './hosting.controller';
 import template from './hosting.html';
 
-import { LOCAL_SEO_FAMILY } from './local-seo/local-seo.constants';
+import { LOCAL_SEO_FAMILY } from '../local-seo/local-seo.constants';
 
 export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('app.hosting', {
