+++ packages/manager/apps/web/client/app/hosting/dashboard/translations/Messages_fr_CA.json
