+++ packages/manager/apps/web/client/app/hosting/dashboard/translations/Messages_pt_PT.json
