+++ packages/manager/apps/web/client/app/hosting/hosting.routing.js
