--- packages/manager/apps/web/client/app/hosting/hostings/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "hostings_title": "Hébergements"
-}
